fix: update schemas and wallets to use Decimal for amounts; fix docker-compose volume path

+++ app/schemas.py
@@ -1,14 +1,16 @@
+from decimal import Decimal
+
 from pydantic import BaseModel, Field, field_validator
 
 
 class OperationRequest(BaseModel):
     wallet_name: str = Field(..., max_length=127)
-    amount: float
+    amount: Decimal
     description: str | None = Field(None, max_length = 256)
     
     @field_validator('amount')
     @classmethod
-    def amount_must_be_positive(cls, v: float) -> float:
+    def amount_must_be_positive(cls, v: Decimal) -> Decimal:
         if v <= 0:
             raise ValueError("Amount must be positive")
         return v
@@ -23,7 +25,7 @@ class OperationRequest(BaseModel):
 
 class CreateWalletRequest(BaseModel):
     name: str = Field(..., max_length=127)
-    initial_balance: float = 0
+    initial_balance: Decimal = 0
 
     @field_validator('name')
     @classmethod
@@ -35,7 +37,7 @@ class CreateWalletRequest(BaseModel):
     
     @field_validator('initial_balance')
     @classmethod
-    def balance_not_negative(cls, v: float) -> float:
+    def balance_not_negative(cls, v: Decimal) -> Decimal:
         if v < 0:
             raise ValueError("balance connot be negative")
         return v
